+++ framework/Furion/JsonSerialization/Extensions/JsonSerializationServiceCollectionExtensions.cs
@@ -1,6 +1,5 @@
 ﻿using Furion.DependencyInjection;
 using Furion.JsonSerialization;
-using Furion.Utilities;
 
 namespace Microsoft.Extensions.DependencyInjection
 {
@@ -32,41 +31,6 @@ namespace Microsoft.Extensions.DependencyInjection
         {
             services.AddSingleton<IJsonSerializerProvider, TJsonSerializerProvider>();
             return services;
-        }
-
-        /// <summary>
-        /// 配置 Json 序列化属性名大写
-        /// </summary>
-        /// <param name="mvcBuilder">Mvc构建器</param>
-        /// <returns>Mvc构建器</returns>
-        public static IMvcBuilder AddJsonSerializerPascalPropertyNaming(this IMvcBuilder mvcBuilder)
-        {
-            // 存储是否启动大写序列化
-            JsonSerializerUtility.EnabledPascalPropertyNaming = true;
-
-            mvcBuilder.AddJsonOptions(options =>
-            {
-                options.JsonSerializerOptions.PropertyNamingPolicy = null;
-            });
-
-            return mvcBuilder;
-        }
-
-        /// <summary>
-        /// 添加时间格式化
-        /// </summary>
-        /// <param name="mvcBuilder"></param>
-        /// <param name="format"></param>
-        /// <returns></returns>
-        public static IMvcBuilder AddDateTimeJsonConverter(this IMvcBuilder mvcBuilder, string format)
-        {
-            mvcBuilder.AddJsonOptions(options =>
-            {
-                options.JsonSerializerOptions.Converters.Add(new DateTimeJsonConverter(format));
-                options.JsonSerializerOptions.Converters.Add(new DateTimeOffsetJsonConverter(format));
-            });
-
-            return mvcBuilder;
         }
     }
 }