+++ framework/Furion/RemoteRequest/Dependencies/IHttpDispatchProxy.cs
@@ -1,9 +1,0 @@
-﻿namespace Furion.RemoteRequest
-{
-    /// <summary>
-    /// Http 代理请求
-    /// </summary>
-    public interface IHttpDispatchProxy
-    {
-    }
-}